introduced MSD and division in subsimulations to improve statistics; the autocorrelations have some scaling issue, the msd might be working but I need a larger input trajectory to check it

+++ src/postprocess/IPCpostprocess.cpp
@@ -6,12 +6,19 @@
 #include <cmath>
 
 
-IPCpostprocess::IPCpostprocess(int inputNumberOfPatches, std::string const& directoryName) {
+IPCpostprocess::IPCpostprocess(int inputNumberOfPatches, size_t const inputNumberOfSubSimulations, std::string const& directoryName) {
+    // input checks
     if (inputNumberOfPatches < 1 || inputNumberOfPatches > 2) {
         std::cerr << "At the moment only 1 and 2 patches are supported.\n";
         exit(1);
     }
     numberOfPatches = inputNumberOfPatches;
+    if (inputNumberOfSubSimulations <= 0) {
+        std::cerr << "The number of subsimulations needs to be a positive integer.\n";
+        exit(1);
+    }
+    numberOfSubSimulations = inputNumberOfSubSimulations;
+
     // clean up old data and recreate output directory
     if(system("rm -rf analysis") != 0) {
         std::cerr << "Unable to delete the old 'analysis/' directory with rm -rf. "
@@ -34,14 +41,13 @@ IPCpostprocess::IPCpostprocess(int inputNumberOfPatches, std::string const& dire
     readOutputFile(directoryName);
 
     // initialize the containers
-    ipcCentersInitialPositions.resize(nIPCs, {0.0, 0.0, 0.0});
     ipcCentersPreviousPositions.resize(nIPCs, {0.0, 0.0, 0.0});
+    ipcCentersCurrentPositions.resize(nIPCs, {0.0, 0.0, 0.0});
     ipcInitialOrientations.resize(nIPCs, {0.0, 0.0, 0.0});
-    ipcPreviousOrientations.resize(nIPCs, {0.0, 0.0, 0.0});
+    ipcCurrentOrientations.resize(nIPCs, {0.0, 0.0, 0.0});
     ipcCentersInitialVelocities.resize(nIPCs, {0.0, 0.0, 0.0});
+    ipcCentersCurrentVelocities.resize(nIPCs, {0.0, 0.0, 0.0});
 
-    meanSquaredDisplacement.resize(simulationDurationInIterations, 0.);
-    totalRotation.resize(simulationDurationInIterations, 0.);
     orientationAutocorrelation.resize(simulationDurationInIterations, 0.);
     velocityAutocorrelation.resize(simulationDurationInIterations, 0.);
 
@@ -55,16 +61,24 @@ IPCpostprocess::IPCpostprocess(int inputNumberOfPatches, std::string const& dire
 }
 
 void IPCpostprocess::run() {
-
-    trajectorySnapshot = 0;
-    readFirstSnapshot();
-
-    for (trajectorySnapshot = 1; trajectorySnapshot <= simulationDurationInIterations; ++trajectorySnapshot) {
-        processSingleSnapshot();
-        std::cout << trajectorySnapshot << " / " << simulationDurationInIterations << "\n";
+    for (double subSymCounter = 0; subSymCounter < numberOfSubSimulations; ++subSymCounter) {
+        for (double subSymSnapshot = 0; subSymSnapshot < subSimulationDuration; ++subSymSnapshot) {
+            const size_t absoluteSnapshot = subSymCounter*subSimulationDuration + subSymSnapshot;
+            readSnapshot(absoluteSnapshot);
+            if (subSymCounter == 0) {
+                updateInitialOrientationAndVelocites();
+                if (subSymCounter == 0) {
+                    updatePreviousPositions(); // for the initial computation of the MSD...
+                }
+            }
+            computeMSD(absoluteSnapshot);
+            computeAutocorrelations(subSymSnapshot);
+            updatePreviousPositions();
+        }
+        std::cout << "Subsym " << subSymCounter << " of " << numberOfSubSimulations << " finished\n";
     }
 
-    printResults();
+    printAutocorrelations(numberOfSubSimulations);
 
     trajectoryFile.close();
     autocorrelationsFile.close();
@@ -72,37 +86,20 @@ void IPCpostprocess::run() {
     //typicalOrientationsFile.close();
 }
 
-void IPCpostprocess::printResults() {
-    // set values that were not computed
-    orientationAutocorrelation[0] = 1.;
-    velocityAutocorrelation[0] = 1.;
-    meanSquaredDisplacement[0] = 0.;
-    totalRotation[0] = 0.;
 
-    double inverseOfNumberOfIPCs = 1./nIPCs;
-    for (size_t i = 0; i <= simulationDurationInIterations; ++i) {
-        autocorrelationsFile << i*printingInterval << "\t"
-                             << orientationAutocorrelation[i]*inverseOfNumberOfIPCs << "\t"
-                             << velocityAutocorrelation[i]*inverseOfNumberOfIPCs << "\n";
-        meanSquaredDisplFile << i*printingInterval << "\t"
-                             << orientationAutocorrelation[i]*inverseOfNumberOfIPCs << "\t"
-                             << velocityAutocorrelation[i]*inverseOfNumberOfIPCs << "\n";
-    }
-}
-
-void IPCpostprocess::readFirstSnapshot() {
-    runConsistencyChecks();
+void IPCpostprocess::readSnapshot(const size_t snapshotNumber) {
+    runConsistencyChecks(snapshotNumber);
 
     for(int i = 0; i < nIPCs; ++i) {
         char uselessChar;
         double useless;
         // IPC center
         trajectoryFile >> uselessChar
-                >> ipcCentersInitialPositions[i][0] >> ipcCentersInitialPositions[i][1] >> ipcCentersInitialPositions[i][2]
-                >> ipcCentersInitialVelocities[i][0] >> ipcCentersInitialVelocities[i][1] >> ipcCentersInitialVelocities[i][2];
+                >> ipcCentersCurrentPositions[i][0] >> ipcCentersCurrentPositions[i][1] >> ipcCentersCurrentPositions[i][2]
+                >> ipcCentersCurrentVelocities[i][0] >> ipcCentersCurrentVelocities[i][1] >> ipcCentersCurrentVelocities[i][2];
         // first patch
         trajectoryFile >> uselessChar
-                >> ipcInitialOrientations[i][0] >> ipcInitialOrientations[i][1] >> ipcInitialOrientations[i][2]
+                >> ipcCurrentOrientations[i][0] >> ipcCurrentOrientations[i][1] >> ipcCurrentOrientations[i][2]
                 >> useless >> useless >> useless;
         // second patch
         if (numberOfPatches == 2) {
@@ -112,60 +109,34 @@ void IPCpostprocess::readFirstSnapshot() {
         }
         for (int j: {0, 1, 2}) {
             // compute orientations
-            ipcInitialOrientations[i][j] -= ipcCentersInitialPositions[i][j];
-            relativePBC(ipcInitialOrientations[i][j]);
-            ipcInitialOrientations[i][j] /= firstPatchEccentricity;
-
-            // set previous state as the current
-            ipcPreviousOrientations[i][j] = ipcInitialOrientations[i][j];
-            ipcCentersPreviousPositions[i][j] = ipcCentersInitialPositions[i][j];
+            ipcCurrentOrientations[i][j] -= ipcCentersCurrentPositions[i][j];
+            relativePBC(ipcCurrentOrientations[i][j]);
+            ipcCurrentOrientations[i][j] /= firstPatchEccentricity;
         }
     }
-
 }
 
-void IPCpostprocess::processSingleSnapshot() {
-    runConsistencyChecks();
-
-    // start reading out IPCs
+void IPCpostprocess::updateInitialOrientationAndVelocites() {
+    // set initial state and previous state as the current
     for(int i = 0; i < nIPCs; ++i) {
-        double ipcCenterPosition [3];
-        double ipcCenterVelocity [3];
-        double ipcOrientation [3];
-        char uselessChar;
-        double useless;
-        // IPC center
-        trajectoryFile >> uselessChar
-                >> ipcCenterPosition[0] >> ipcCenterPosition[1] >> ipcCenterPosition[2]
-                >> ipcCenterVelocity[0] >> ipcCenterVelocity[1] >> ipcCenterVelocity[2];
-        // first patch
-        trajectoryFile >> uselessChar
-                >> ipcOrientation[0] >> ipcOrientation[1] >> ipcOrientation[2]
-                >> useless >> useless >> useless;
-        // second patch
-        if (numberOfPatches == 2) {
-            trajectoryFile >> uselessChar;
-            trajectoryFile.ignore(500, '\n');
-        }
-        // compute orientation
         for (int j: {0, 1, 2}) {
-            ipcOrientation[j] -= ipcCenterPosition[j];
-            relativePBC(ipcOrientation[j]);
-            ipcOrientation[j] /= firstPatchEccentricity;
-        }
-
-        for (int j: {0, 1, 2}) {
-            // compute mean squared displacement and total rotation
-            //meanSquaredDisp +=
-
-            // compute autocorrelations
-            orientationAutocorrelation[trajectorySnapshot] += ipcOrientation[j]*ipcInitialOrientations[i][j];
-            velocityAutocorrelation[trajectorySnapshot] += ipcCenterVelocity[j]*ipcCentersInitialVelocities[i][j];
+            ipcInitialOrientations[i][j] = ipcCurrentOrientations[i][j];
+            ipcCentersInitialVelocities[i][j] = ipcCentersCurrentVelocities[i][j];
         }
     }
 }
 
-void IPCpostprocess::runConsistencyChecks() {
+void IPCpostprocess::updatePreviousPositions() {
+    // set initial state and previous state as the current
+    for(int i = 0; i < nIPCs; ++i) {
+        for (int j: {0, 1, 2}) {
+            ipcCentersPreviousPositions[i][j] = ipcCentersCurrentPositions[i][j];
+        }
+    }
+
+}
+
+void IPCpostprocess::runConsistencyChecks(size_t const snapshotNumber) {
     int nIPCsCheck;
     double simulationBoxSideCheck;
     double time;
@@ -180,9 +151,9 @@ void IPCpostprocess::runConsistencyChecks() {
                   << simulationBoxSide << " != " << simulationBoxSideCheck << ".\n";
         exit(1);
     }
-    if (trajectorySnapshot*printingInterval != time) {
+    if (snapshotNumber*printingInterval != time) {
         std::cerr << "Inconsistency in the time.\n"
-                  << trajectorySnapshot*printingInterval << " != " << time << ".\n";
+                  << snapshotNumber*printingInterval << " != " << time << ".\n";
         exit(1);
     }
 }
@@ -213,7 +184,41 @@ void IPCpostprocess::readOutputFile(const std::string &directoryName) {
     // process data
     simulationBoxSide = std::cbrt(nIPCs/density);
     simulationDurationInIterations = (size_t)simulationTotalDuration/printingInterval;
+    subSimulationDuration = (size_t)simulationDurationInIterations/numberOfSubSimulations;
 
     // scale lenghts
     firstPatchEccentricity /= simulationBoxSide;
+}
+
+void IPCpostprocess::computeMSD(size_t const snapshotNumber) {
+    double MSD = 0.;
+    for(int i = 0; i < nIPCs; ++i) {
+        for (int j: {0, 1, 2}) {
+            double delta_xj = ipcCentersCurrentPositions[i][j] - ipcCentersPreviousPositions[i][j];
+            relativePBC(delta_xj);
+            MSD += std::pow(delta_xj,2);
+        }
+    }
+    meanSquaredDisplFile << snapshotNumber << "\t" << MSD << "\n";
+}
+
+void IPCpostprocess::computeAutocorrelations(size_t const snapshotNumber) {
+
+    for(int i = 0; i < nIPCs; ++i) {
+        for (int j: {0, 1, 2}) {
+            // compute autocorrelations
+            orientationAutocorrelation[snapshotNumber] += ipcCurrentOrientations[i][j]*ipcInitialOrientations[i][j];
+            velocityAutocorrelation[snapshotNumber] += ipcCentersCurrentVelocities[i][j]*ipcCentersInitialVelocities[i][j];
+        }
+    }
+}
+
+void IPCpostprocess::printAutocorrelations(size_t const numberOfSubSimulations) {
+
+    double scalingFactor = 1./nIPCs/numberOfSubSimulations;
+    for (size_t i = 0; i <= simulationDurationInIterations; ++i) {
+        autocorrelationsFile << i*printingInterval << "\t"
+                             << orientationAutocorrelation[i]*scalingFactor << "\t"
+                             << velocityAutocorrelation[i]*scalingFactor << "\n";
+    }
 }
